Add files via upload
The website includes : 
1. Homepage: Create an engaging landing page that showcases the key features and benefits
of the online recruitment solution, emphasizing its efficiency for hiring managers and HR
professionals.
2. User Login and Register Pages: Develop secure login and registration pages specifically
tailored for companies. This will enable them to access the recruitment solution and
utilize its features effectively.
3. Candidate Search: Implement a search functionality that allows users to find candidates
based on location and job roles. Additionally, design a section where the fetched
candidates can be listed for later ranking and evaluation